+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.1
+
+- File Statistics to count of total files, directories, and Dart files. 
+- TODO and FIXME Comments to scan files for TODO and FIXME comments and list them in a separate collapsible section.
+- Dependency Analysis to list all external package dependencies used in the project.
+- Code Metrics to calculate and display simple code metrics like lines of code, comment percentage, etc.
+
 ## 1.0.0
 
 - Initial version.